finish refactor of Lexical analyses
